+++ src/App.js
@@ -115,8 +115,8 @@ const LoadableSonarFeedPage = Loadable({
   loading: () => <PageLoader />
 })
 
-const LoadableAssetsOverviewPage = Loadable({
-  loader: () => import('./pages/assets/AssetsOverviewPage'),
+const LoadableWatchlistsPage = Loadable({
+  loader: () => import('./pages/Watchlists'),
   loading: () => <PageLoader />
 })
 
@@ -277,7 +277,7 @@ export const App = ({
             <CreateAccountFreeTrial {...props} isLoggedIn={isLoggedIn} />
           )}
         />
-        <Route exact path='/assets' component={LoadableAssetsOverviewPage} />
+        <Route exact path='/assets' component={LoadableWatchlistsPage} />
         <Route
           exact
           path='/watchlists'
