+++ src/pages/Marketing/MarketingPage.js
@@ -13,7 +13,7 @@ import MobileHeader from '../../components/MobileHeader/MobileHeader'
 import { MobileOnly } from '../../components/Responsive'
 import IndexTab from './IndexTabs/IndexTab'
 import WatchlistCards from '../../components/Watchlists/WatchlistCards'
-import { BASIC_CATEGORIES } from '../assets/assets-overview-constants'
+import { BASIC_CATEGORIES } from '../../ducks/Watchlists/utils'
 import MyWatchlist from '../../components/Watchlists/MyWatchlist'
 import { PATHS } from '../../App'
 import AlphaBlock from './AlphaBlock/AlphaBlock'
